+++ tests/ui/no_comma_after_zeroable.rs
@@ -1,0 +1,10 @@
+use dst_factory::make_dst_factory;
+
+#[make_dst_factory(create, zeroable no_std)]
+struct NoCommaAfterZeroable {
+    id: i32,
+    data: [u8],
+}
+
+fn main() {
+}