+++ tests/ui/zeroable_str.rs
@@ -1,0 +1,10 @@
+use dst_factory::make_dst_factory;
+
+#[make_dst_factory(zeroable)]
+struct ZeroableStr {
+    id: i32,
+    data: str,
+}
+
+fn main() {
+}